+++ gum-gum/CHANGES.md
@@ -1,4 +1,7 @@
 
+# 2.3.0
+- Adds bf, irisid, and iriscat params
+- Updates bidUnitInCents and removes multiplication of bidprice
 # 2.2.0
 - Adds support for CCPA (US Privacy)
 # 2.1.0